DataEnabledControl

Documentation for DataEnabledControl.

DataEnabledControl

Install

Quick start

Functions and Types

Background: data-driven control

Install

In the Julia REPL package mode:

pkg> add DataEnabledControl

Quick start

For a quick start please check example.ipynb notebook under docs.

Functions and Types

DataEnabledControl.DataEnabledControl

DataEnabledControl.ConstraintMatrices

DataEnabledControl.DeePC_struct

DataEnabledControl.build_data_matrix

DataEnabledControl.hankellize

DataEnabledControl.run_DeePC

DataEnabledControl.DataEnabledControl — Module

DataEnabledControl.jl is a Julia package implementing the DeePC algorithm in the paper "Coulson, J., Lygeros, J., & Dörfler, F. (2019, June). Data-enabled predictive control: In the shallows of the DeePC. In 2019 18th European Control Conference (ECC) (pp. 307-312). IEEE."

DataEnabledControl.ConstraintMatrices — Type

ConstraintMatrices

Stores the constraint matrices representing polyhedron constraints of the form: Ay * y <= by and Au * u <= bu.

DataEnabledControl.DeePC_struct — Type

DeePC_struct

Used to formulate the DeePC optimization problem Q: the weighting matrix of the output signal R: the weighting matrix of the input signal PastDataMatrix: the left tall matrix [Uₚ; Yₚ] in the paper FutureInputMatrix: Uf according to the paper FutureOutputMatrix: Yf according to the paper Constrained: true if the system is input or output constrained constraintMatrices: params of polyhedral constraints

DataEnabledControl.build_data_matrix — Method

build_data_matrix(U::Matrix{Float64}, Y::Matrix{Float64}, Tᵢₙᵢ::Int, N::Int)

Takes the data collected U,Y through a persistently excited experiment and returns the matrices [Up; Yp], Uf, Yf used in the construction of the DeePC algorithm.

DataEnabledControl.hankellize — Method

hankellize(U::Matrix{Float64}, L::Int)

Constructs the hankel matrix Hof a matrixU, withLnumber of shifts.Hhere corresponds toH_L(U)` in the DeePC paper.

DataEnabledControl.run_DeePC — Method

run_DeePC(prob::DeePC_struct, Uᵢₙᵢ::Vector{Float64}, Yᵢₙᵢ::Vector{Float64}; slack_var = false, λy = 1.0, λg = 0.0)

Constructs and solves the DeePC optimization problem, with a slack variable or without (noisy/nonlinear or not). It returns uₜ and yₜ, the input and output sequence solutions of the DeePC over the N-horizon. The regularization weights (the λ's) are set to zero and the slack variable is deactivated in the default settings.